refactor src, delete unused assets, add coverage todo: more test and increase coverage
