+++ .nycrc.json
@@ -1,0 +1,13 @@
+{
+  "all": true,
+  "include": "src",
+  "check-coverage": true,
+  "per-file": true,
+  "lines": 99,
+  "statements": 99,
+  "functions": 99,
+  "branches": 99,
+  "exclude" : [
+      "src/public/**/*.*"
+    ]
+}
