+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "start": "node src/Index.js",
     "dev": "supervisor -e 'js|hbs' src/Index.js",
-    "test": "mocha 'test/**/*.js'"
+    "test": "mocha 'test/**/*.js'",
+    "coverage": "nyc --reporter=html --reporter=text --reporter=json-summary mocha 'test/**/*.js'"
   },
   "license": "MIT",
   "dependencies": {
@@ -26,9 +27,13 @@
     "os": "^0.1.2",
     "request": "2.88.0",
     "supervisor": "0.12.0",
-    "uuid": "3.3.3",
+    "uuid": "3.3.3"
+  },
+  "devDependencies": {
     "chai": "4.2.0",
-    "mocha": "9.1.3"
+    "coverage-badge-creator": "^1.0.8",
+    "mocha": "9.1.3",
+    "nyc": "^15.1.0"
   },
   "repository": {
     "type": "git",
