+++ src/public/pages/application-variable/main.js
