+++ src/views/partials/scripts.hbs
@@ -35,7 +35,6 @@
 
 <!-- Custom Theme JavaScript -->
 <script src="/dist/js/sb-admin-2.js"></script>
-<script src="/dist/js/custom/app.js"></script>
 
 <script type="text/javascript">
 	$(document).ready(function($) {
@@ -46,5 +45,5 @@
 </script>
 
 {{#if javascriptModule}}
-	<script src="/dist/js/custom/{{javascriptModule}}/main.js"></script>
+	<script src="/pages/{{javascriptModule}}/main.js"></script>
 {{/if}}
